# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenTogg API",
        "description": "Time tracking REST API. Authenticate with a Bearer token created in Settings.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/time-entries": {
            "get": {
                "tags": [
                    "Time Entries"
                ],
                "summary": "List completed time entries",
                "operationId": "678173875d6a18f465e1c010fd87f2b0",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Filter from date (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "description": "Filter to date (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (1-100, default 50)",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated time entries"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Time Entries"
                ],
                "summary": "Start a new timer",
                "description": "Creates a new time entry and starts the timer. Stops any currently running timer.",
                "operationId": "dd393a1b73800a04e0276e2700b67da9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Working on feature X",
                                        "nullable": true
                                    },
                                    "vector_id": {
                                        "description": "Vector external_id",
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Timer started"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/time-entries/{id}": {
            "get": {
                "tags": [
                    "Time Entries"
                ],
                "summary": "Get a time entry",
                "operationId": "9e5f5fef27fd43dc2844b28c7aef100b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Time entry external_id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Time entry details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Time Entries"
                ],
                "summary": "Update a time entry",
                "operationId": "6faac3614331e14199c56e544ad8b383",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Time entry external_id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "vector_id": {
                                        "description": "Vector external_id",
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "tag_ids": {
                                        "description": "Tag external_ids",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Time entry updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Time Entries"
                ],
                "summary": "Delete a time entry",
                "operationId": "d32390c976b1223acef9bfcabe615e73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Time entry external_id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Time entry deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/time-entries/current": {
            "get": {
                "tags": [
                    "Time Entries"
                ],
                "summary": "Get the currently running timer",
                "operationId": "e409b4947153e253213bc2767acfa525",
                "responses": {
                    "200": {
                        "description": "Current timer status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "running": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/time-entries/current/stop": {
            "patch": {
                "tags": [
                    "Time Entries"
                ],
                "summary": "Stop the running timer",
                "operationId": "4f42cce29dca6456866a862f6b43bb62",
                "responses": {
                    "200": {
                        "description": "Timer stopped"
                    },
                    "404": {
                        "description": "No running timer"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}